unfeat: remove list results display and keep only table display
