+++ itou/templates/apply/includes/job_applications_filters/offcanvas_body.html
@@ -1,6 +1,5 @@
 <div class="offcanvas-body" id="offcanvasApplyFiltersContent"{% if request.htmx %} hx-swap-oob="true"{% endif %}>
     {% include "apply/includes/job_applications_filters/status.html" %}
-    {% if display_kind %}<input id="display-kind" type="hidden" name="display" value="{{ display_kind }}">{% endif %}
     <input id="id_order" type="hidden" name="order" value="{{ order }}">
     {% if job_applications_list_kind is JobApplicationsListKind.SENT or job_applications_list_kind is JobApplicationsListKind.RECEIVED %}
         {% include "apply/includes/job_applications_filters/selected_jobs.html" %}
