+++ itou/templates/apply/includes/list_reset_filters.html
@@ -10,9 +10,7 @@
 {% if btn_dropdown_filter|default:False %}
     <div class="ms-md-auto" id="apply-list-filter-counter"{% if request.htmx %} hx-swap-oob="true"{% endif %}>
         {% if filters_counter > 0 %}
-            <a href="{{ reset_url }}?display={{ display_kind }}&order={{ order }}"
-               class="btn btn-ico btn-dropdown-filter"
-               aria-label="Réinitialiser {{ filters_counter|pluralizefr:"le filtre actif,les filtres actifs" }}">
+            <a href="{{ reset_url }}?order={{ order }}" class="btn btn-ico btn-dropdown-filter" aria-label="Réinitialiser {{ filters_counter|pluralizefr:"le filtre actif,les filtres actifs" }}">
                 <i class="ri-eraser-line fw-medium" aria-hidden="true"></i>
                 <span>Effacer tout</span>
             </a>
@@ -20,9 +18,7 @@
     </div>
 {% else %}
     {% if filters_counter > 0 %}
-        <a href="{{ reset_url }}?display={{ display_kind }}&order={{ order }}"
-           class="btn btn-ico btn-block btn-outline-primary"
-           aria-label="Réinitialiser {{ filters_counter|pluralizefr:"le filtre actif,les filtres actifs" }}">
+        <a href="{{ reset_url }}?order={{ order }}" class="btn btn-ico btn-block btn-outline-primary" aria-label="Réinitialiser {{ filters_counter|pluralizefr:"le filtre actif,les filtres actifs" }}">
             <i class="ri-eraser-line fw-medium" aria-hidden="true"></i>
             <span>Effacer tout</span>
         </a>
