+++ itou/templates/apply/list_for_job_seeker.html
@@ -21,9 +21,6 @@
                 <div class="col-12">
                     <div class="d-flex flex-column flex-md-row align-items-md-center mb-3 mb-md-4">
                         <div class="flex-md-grow-1">{% include "apply/includes/list_counter.html" %}</div>
-                        <div class="d-flex gap-2 mt-3 mt-md-0 ms-0 ms-md-2" role="group" aria-label="Changement du mode d'affichage des candidatures">
-                            {% include "common/list_view_switcher.html" with display_kind=display_kind request=request only %}
-                        </div>
                     </div>
                     {% include "apply/includes/list_job_applications.html" with order=order %}
                 </div>
