+++ itou/templates/apply/list_for_siae.html
@@ -44,11 +44,8 @@
                         <div class="flex-column flex-md-row mt-3 mt-md-0">
                             {% bootstrap_field filters_form.job_seeker wrapper_class="w-lg-400px" show_label=False %}
                         </div>
-                        <div class="d-flex gap-2 mt-3 mt-md-0 ms-0 ms-md-2" role="group" aria-label="Changement du mode d'affichage des candidatures">
-                            {% include "common/list_view_switcher.html" with display_kind=display_kind request=request only %}
-                        </div>
                     </div>
-                    {% include "apply/includes/list_job_applications.html" with job_applications_page=job_applications_page filters_counter=filters_counter request=request csrf_token=csrf_token pending_states_job_applications_count=pending_states_job_applications_count list_exports_url=list_exports_url job_applications_list_kind=job_applications_list_kind JobApplicationsListKind=JobApplicationsListKind display_kind=display_kind order=order ITOU_HELP_CENTER_URL=ITOU_HELP_CENTER_URL only %}
+                    {% include "apply/includes/list_job_applications.html" with job_applications_page=job_applications_page filters_counter=filters_counter request=request csrf_token=csrf_token pending_states_job_applications_count=pending_states_job_applications_count list_exports_url=list_exports_url job_applications_list_kind=job_applications_list_kind JobApplicationsListKind=JobApplicationsListKind order=order ITOU_HELP_CENTER_URL=ITOU_HELP_CENTER_URL only %}
                 </div>
             </div>
         </div>
