+++ itou/templates/approvals/includes/list_results.html
@@ -6,13 +6,7 @@
             </div>
         </div>
     {% else %}
-        {% if display_kind.is_table %}
-            <div class="table-responsive mt-3 mt-md-4">{% include "approvals/includes/list_table.html" %}</div>
-        {% elif display_kind.is_list %}
-            {% for approval in approval_list %}
-                {% include "approvals/includes/list_card.html" with approval=approval %}
-            {% endfor %}
-        {% endif %}
+        <div class="table-responsive mt-3 mt-md-4">{% include "approvals/includes/list_table.html" %}</div>
         {% include "includes/pagination.html" with page=page_obj boost=True boost_target="#approvals-list" boost_indicator="#approvals-list" request=request only %}
     {% endif %}
 </div>
@@ -20,5 +14,4 @@
 {% if request.htmx %}
     {% include "approvals/includes/list_counter.html" %}
     {% include "approvals/includes/list_reset_filters.html" %}
-    {% include "common/list_view_switcher.html" with display_kind=display_kind request=request only %}
 {% endif %}
