+++ itou/templates/approvals/list.html
@@ -49,7 +49,6 @@
                               hx-swap="outerHTML"
                               hx-push-url="true"
                               hx-include="#id_job_seeker">
-                            <input id="display-kind" type="hidden" name="display" value="{{ display_kind }}">
                             <div class="btn-dropdown-filter-group mb-3 mb-md-4">
                                 {% include "approvals/includes/approvals_filters/status.html" %}
                                 {% include "includes/btn_dropdown_filter/radio.html" with field=filters_form.expiry only %}
@@ -62,9 +61,6 @@
                             <div class="flex-md-grow-1">{% include "approvals/includes/list_counter.html" %}</div>
                             <div class="flex-column flex-md-row mt-3 mt-md-0">
                                 {% bootstrap_field filters_form.job_seeker wrapper_class="w-lg-400px" show_label=False %}
-                            </div>
-                            <div class="d-flex gap-2 mt-3 mt-md-0 ms-0 ms-md-2" role="group" aria-label="Changement du mode d'affichage des salariés et PASS IAE">
-                                {% include "common/list_view_switcher.html" with display_kind=display_kind request=request only %}
                             </div>
                         </div>
                         {% include "approvals/includes/list_results.html" %}
