--- itou/templates/common/list_view_switcher.html
@@ -1,16 +1,0 @@
-<button id="table_display_button"
-        class="btn btn-ico btn-block {{ display_kind.is_table|yesno:'btn-secondary,btn-link' }}"
-        data-emplois-setter-target="#display-kind"
-        data-emplois-setter-value="{{ display_kind.TABLE }}"
-        {% if request.htmx %}hx-swap-oob="true"{% endif %}>
-    <i class="ri-table-line" aria-hidden="true"></i>
-    <span>Tableau</span>
-</button>
-<button id="list_display_button"
-        class="btn btn-ico btn-block {{ display_kind.is_list|yesno:'btn-secondary,btn-link' }}"
-        data-emplois-setter-target="#display-kind"
-        data-emplois-setter-value="{{ display_kind.LIST }}"
-        {% if request.htmx %}hx-swap-oob="true"{% endif %}>
-    <i class="ri-stacked-view" aria-hidden="true"></i>
-    <span>Liste</span>
-</button>
